gcc-configure-sdk.inc: Ensure target sysroot is under $prefix
Currently `gcc -print-sysroot` outputs the literal --with-sysroot
setting:

$ /mnt/rtollert/sdk1/sysroots/x86_64-nilrtsdk-linux/usr/bin/armv7a-vfp-neon-nilrt-linux-gnueabi/arm-nilrt-linux-gnueabi-gcc -print-sysroot
/usr/local/oecore-x86_64/sysroots/armv7a-vfp-neon-nilrt-linux-gnueabi

However, if the default sysroot (--with-sysroot=) is underneath the
prefix (--prefix=), then gcc rewrites the sysroot at runtime based on
the actual location of gcc. Allowing this to happen obviates the need to
supply a sysroot on the command line in many situations.

$ /mnt/rtollert/sdk2/sysroots/x86_64-nilrtsdk-linux/usr/bin/armv7a-vfp-neon-nilrt-linux-gnueabi/arm-nilrt-linux-gnueabi-gcc -print-sysroot
/mnt/rtollert/sdk2/sysroots/x86_64-nilrtsdk-linux/usr/bin/armv7a-vfp-neon-nilrt-linux-gnueabi/../../../../armv7a-vfp-neon-nilrt-linux-gnueabi
$ /mnt/rtollert/sdk2/sysroots/x86_64-nilrtsdk-linux/usr/bin/armv7a-vfp-neon-nilrt-linux-gnueabi/arm-nilrt-linux-gnueabi-gcc -v |& sed -e 's/ /\n/g' | egrep -- '--prefix|--with-sysroot'
--prefix=/usr/local/oecore-x86_64/sysroots
--with-sysroot=/usr/local/oecore-x86_64/sysroots/armv7a-vfp-neon-nilrt-linux-gnueabi

cf gcc/gcc.c:process_command() in the
defined(TARGET_SYSTEM_ROOT_RELOCATABLE), and
libiberty/make-relative-prefix.c, in the gcc sources.

Implement this by setting the prefix one directory up, to be
../sysroots/.

Natinst-ReviewBoard-ID: 52189
Acked-by: Gratian Crisan <gratian.crisan@ni.com>
Acked-by: David Curtiss <david.curtiss@ni.com>

Signed-off-by: Richard Tollerton <rich.tollerton@ni.com>