elfutils: clean up patches
In the upgrade a large number of Upstream-Status tags were dropped, so add them
back.  I'm taking the stand that copying a patch Debian is carrying doesn't
count as a backport.

Remove two Debian-specific patches (one for Hurd, one for kfreebsd) so
we're not carrying useless patches.

Remove two patches that are no longer applied.

Signed-off-by: Ross Burton <ross.burton@intel.com>