gcc-runtime: Add missing libc dependency
For reasons lost in the depths of time, perhaps performane related,
we only have a dependency on libc at packaging time. This is too late,
as demonstrated by a recent build failure on non-IA builds where
the glibc 2.29 upgrade had been removed from the build:

ld: recipe-sysroot/usr/lib/../lib/libstdc++.so: undefined reference to `log@GLIBC_2.29'

libstdc++ should have been rebuilt but had not as the dependency
wasn't present.

Add the missing dependency to avoid this problem (and drop the other
form of dependency which is no longer needed).

Signed-off-by: Richard Purdie <richard.purdie@linuxfoundation.org>