kernel-module-split: Identify kernel modconf files as configuration files
Currently the modconf fragments representing the configuration for
kernel modules are written out to appropriate .conf files and added to
the FILES variable. However they are not identified as 'configuration
files' and installing a new version of a kernel module results in a
conflict and a failed installed (because of the respective .conf file
is already in place from a previous install).

Add the generated .conf files to the CONFFILES variable denoting their
true nature. This allows proper handling from the package management
system.

Signed-off-by: Gratian Crisan <gratian.crisan@ni.com>