openssl: upgrade to 1.0.2u
CVEs (there is nothing deliberate in this list, I'm just reading off the
changelog):

CVE-2019-1547
CVE-2019-1563
CVE-2019-1552
CVE-2019-1559
CVE-2018-5407
CVE-2018-0734
CVE-2018-0732
CVE-2018-0737

You are probably wondering why we're not using 1.1 yet. The reason is that
LabVIEW is presently dlsym()ing into openssl like a fool, and 1.1 broke that. At
present, it is believed that the only thing we're missing out on is
Racoon (CVE-2020-1968).

Signed-off-by: Richard Tollerton <rich.tollerton@ni.com>