opkg-utils: split package into shell-only/python-only scripts
Not all scripts in opkg-utils require python3 as a dependency; some of
them are implemented in shell and only require that. This commit splits
the opkg-utils package into two subpackages so that individual scripts
can be included without requiring all scripts, particularly the ones
that require a python3 runtime.

For compatibility, the opkg-utils package itself continues to exist as
a metapackage that depends on all its subpackages so that installing
"opkg-utils" would install all the things it would before.

Signed-off-by: Brandon Streiff <brandon.streiff@ni.com>