Revert "bitbake.conf: drop pod2man from hosttools"
This reverts commit c2f77fc480bf2ae0f20208e48b5f0f714d4be0d3.

My presumption that removing pod2man from the hosttools would be
unobtrusive was not correct. pod2man is still required for the
opkg-utils build - and I'm not sure how it works upstream.

In the mean time, revert this commit, since it was only ever an sstate
cache efficiency optimization and it breaks the build as-is.

Signed-off-by: Alex Stewart <alex.stewart@ni.com>