Revert "bootlogd: Allow console writes to ttyPS*"
This reverts commit 7b423fe4fd23bc3cd89d18ea7a712d324a224b8a.

The revert-target commit applies a patch to bootlogd to include
Xilinx-zynq UART consoles to the list of monitored devices for bootlogd.
Sysvinit is ancient and being phased out by OE-core. Given that upstream
is unlikely to ever accept this patch upstream, it should not live in
our OE-core branch. And given that we are not supporting ARM on
hardknott at this time, it is most appropriate to remove the .patch
entirely.

Signed-off-by: Alex Stewart <alex.stewart@ni.com>