elfutils: Fix ptest dependencies
The elfutils ptests require debug symbols for the libc to be available, else
we see failures such as those on the autobuilder for the fast ptest image
on arm (the dbg symbols are pulled in by other recipes in other images).

Also fix various test skips/error messages due to missing gcc/ld and
development headers.

Signed-off-by: Richard Purdie <richard.purdie@linuxfoundation.org>