insane.bbclass: use _${PN} to namespace native-last
The `native-last` QA check verifies that recipes which inherit the
`native` bbclass do so at the end of their recipes. The check is
supposed to be skippable via the INSANE_SKIP mechanism, but its current
implementation does not comprehend the usual _${PN} variable
namespacing, and instead checks `INSANE_SKIP` only.

Instead, let bitbake expand `INSANE_SKIP_${PN}`, so that the check can
be skipped using the same variable as all other skips.

Signed-off-by: Alex Stewart <alex.stewart@ni.com>