lttng-modules: Make it build when CONFIG_TRACEPOINTS is not enabled again
Makefile-Do-not-fail-if-CONFIG_TRACEPOINTS-is-not-en.patch was updated
with the 2.13.0 update (commit 5dad15af), but unfortunately it no
longer did what it was intended to do. There is also
0001-src-Kbuild-change-missing-CONFIG_TRACEPOINTS-to-warn.patch, which
was based on the former patch, but it too does not solve the problem.

Unify the two patches, and actually disable building of the kernel
module if CONFIG_TRACEPOINTS is not enabled, the way it was intended.

Signed-off-by: Peter Kjellerstedt <peter.kjellerstedt@axis.com>
Signed-off-by: Richard Purdie <richard.purdie@linuxfoundation.org>