scripts: Add a conversion script to use SPDX license names
We're standardising on SPDX license names so it is overdue to change
the ones in the LICENSE fileds to the SPDX values. Add a conversion
script which makes this straightforward on the most part.

Ultimately this allows the core code to be more efficient and not need
to consult the mappings.

Signed-off-by: Richard Purdie <richard.purdie@linuxfoundation.org>
(cherry picked from commit 512cd4ca91bc5107d68b7e721257a8f62f878994)
Signed-off-by: Mike Petersen <mike.petersen@ni.com>