libc-package: Fix output name of localedef
locale packages with different encoding but same locale name run
localedef with same name in their postinst. So if more than one of those
packages is installed, the last running postinst will override the
charmap for previously installed locales.

So change the output path to be the same as package name.

Signed-off-by: Chaitanya Vadrevu <chaitanya.vadrevu@ni.com>