package-index: Convert package-index into a bbclass.
Currently, package-index is a bit strange because it's a
bitbake recipe that implements a task. That task is
already used elsewhere and could be leveraged further
for BUILD_IMAGES_FROM_FEEDS.

This change moves the logic for do_package_index into its
own bbclass and has package-index.bb inherit from there.

Signed-off-by: Charlie Johnston <charlie.johnston@ni.com>