packagefeed: Add new oe.packagefeed classes
This change adds the infrastructure to build a
packagefeed to oe. Specifically, packagefeed.py
adds an abstract class and the logic to create a
hardlinked feed directory while the package type
specific classes allow the proper package indexer,
deploy directory, etc to be selected for a build.

The logic to construct the package feed is modified
from similar logic used in oe.package_manager when
constructing a feed for rootfs builds with some minor
differences to allow filtering out packages which may
already be in other feeds.

Signed-off-by: Charlie Johnston <charlie.johnston@ni.com>