packagefeed.bbclass: Add new bbclass for building feeds.
Add a new bbclass that allows building a feed using the
new oe.packagefeed class. This new bbclass inherits
from packagegroup so that there is also a package in the
feed that represents all packages in the feed.

The variable FEED_DEPENDS can be used to specify a feed
that the packagefeed depends on and will be available
side-by-side. This prevents duplicate packages in the
two feeds.

Additionally, there are packagefeed_<type> bbclasses to
define package type specific configurations.

Signed-off-by: Charlie Johnston <charlie.johnston@ni.com>